+++ evennia/server/amp_client.py
@@ -5,9 +5,11 @@ Portal. This module sets up the Client-side communication.
 """
 
 import os
+from django.conf import settings
 from evennia.server.portal import amp
 from twisted.internet import protocol
 from evennia.utils import logger
+from evennia.utils.utils import class_from_module
 
 
 class AMPClientFactory(protocol.ReconnectingClientFactory):
@@ -33,7 +35,7 @@ class AMPClientFactory(protocol.ReconnectingClientFactory):
 
         """
         self.server = server
-        self.protocol = AMPServerClientProtocol
+        self.protocol = class_from_module(settings.AMP_CLIENT_PROTOCOL_CLASS)
         self.maxDelay = 10
         # not really used unless connecting to multiple servers, but
         # avoids having to check for its existence on the protocol